Create Dealer class
This had a major impact on the game. The flow is now affected
differently with each class given responsibility. The dealer is in
charge of handling the deck and controlling payouts. The hand holds the
cards and the bet. The player controls the player bank. Script still
controls flow of the game.